+++ bob-apps/bin/config_cluster_generator/mod.rs
@@ -1,2 +1,3 @@
 pub mod center;
+pub mod pattern;
 pub mod utils;